[COMMENTS] update model, services and controllers for the comment listing

+++ www/Controllers/Comments.php
@@ -4,18 +4,41 @@ namespace App\Controllers;
 
 use App\Core\Validator;
 use App\Models\Comment;
-use App\Models\User;
 use App\Core\Controller;
 use App\Errors\NotFoundError;
+use App\Errors\Unauthorized;
 use App\Errors\ValidatorError;
+use App\Policies\CommentPolicy;
+use App\Services\CommentServices;
 
 class Comments extends Controller
 {
 
     public function index()
-    {
-        $comments = Comment::all();
-        return $comments;
+    {   
+        $query = request()->getQuery();
+
+        if($query->has('page_id')){
+           
+            $comments = Comment::findMany([
+                'page_id'=> (int) $query->get('page_id'),
+                'status'=> 'validated',
+                'comment_id'=> null
+            ]);
+            
+            $comments = $comments->map(function($child){
+                return CommentServices::getCommentConversation($child);
+            });
+
+            return $comments;
+        }else{
+            $authUser = request()->auth()->user();
+            if(!$authUser) throw new Unauthorized();
+            CommentPolicy::index($authUser);
+            $comments = Comment::all();
+            return $comments;
+        }
+
     }
 
     public function show($params)
@@ -23,10 +46,9 @@ class Comments extends Controller
         $comment = Comment::fetch($params['id']);
         if (!$comment) throw new NotFoundError();
 
-        // $comment->getAuthorData();
-        // var_dump($comment);
-        // $author = $comment->getAuthorId();
-        
+        $comment->author = $comment->getAuthor();
+        $comment->comment = $comment->getComment();
+        $comment->comments = $comment->getComments();
 
         return $comment;
     }
@@ -35,11 +57,11 @@ class Comments extends Controller
     {
         $payload = request()->json();
 
-        $validator = new Validator($payload, [
+        $validator = new Validator();
+        $validator->validate($payload, [
             'content' => 'required',
             'author_id' => 'required',
-            'article_id' => 'required',
-            'comment_id' => 'required',
+            'page_id' => 'required',
         ]);
 
         if ($validator->hasErrors()) {
@@ -50,7 +72,7 @@ class Comments extends Controller
 
         $comment->setContent($payload['content']);
         $comment->setAuthorId($payload['author_id']);
-        $comment->setArticleId($payload['article_id']);
+        $comment->setPageId($payload['page_id']);
 
         if (isset($payload['comment_id']))
             $comment->setCommentId($payload['comment_id']);
@@ -77,10 +99,20 @@ class Comments extends Controller
     {
 
         $comment = Comment::fetch($params['id']);
-
         if (!$comment) throw new NotFoundError();
+        
+        $answersCount = Comment::findMany([
+            'comment_id' => $params['id']
+        ])->count();
 
-        $comment->destroy();
+
+        if($answersCount > 0){
+            $comment->setContent("Ce commentaire a été supprimé");
+
+            $comment->save();
+        }else{
+            $comment->destroy();
+        }
 
         return $comment;
     }
